# docs/swagger.yaml
basePath: /api
definitions:
...
  lab_internal_app_DTO.ChangeSolarPanelAreaRequest:
    properties:
      area:
        type: number
    required:
    - area
    type: object
...
  lab_internal_app_ds.RequestPanels:
    properties:
      area:
        type: number
      solar_panel_id:
        type: integer
      solar_panel_request_id:
        type: integer
      solarPanel:
        $ref: '#/definitions/lab_internal_app_ds.SolarPanel'
    type: object
  lab_internal_app_ds.SolarPanel:
    properties:
      depth:
        type: integer
      description:
        type: string
      efficiency:
        type: string
      height:
        type: integer
      id:
        type: integer
      image:
        type: string
      isDelete:
        type: boolean
      power:
        type: number
      title:
        type: string
      type:
        type: string
      width:
        type: integer
    type: object
...
host: localhost:8001
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: API для управления солнечными панелями и расчета их мощности
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Solar Panels Power Calculator API
  version: "1.0"
paths:
...
  /solarpanel-requests/{id}/{solarpanelId}:
    delete:
      consumes:
      - application/json
      description: Удаляет солнечную панель из заявки пользователя
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: ID солнечной панели
        in: path
        name: solarpanelId
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Панель успешно удалена
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Заявка недоступна / Панель отсутствует
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка или панель не найдены
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Удалить панель из заявки
      tags:
      - RequestPanels
    put:
      consumes:
      - application/json
      description: Обновляет значение площади для панели в заявке
      parameters:
      - description: ID заявки
        in: path
        name: id
        required: true
        type: integer
      - description: ID солнечной панели
        in: path
        name: solarpanelId
        required: true
        type: integer
      - description: Новое значение площади
        in: body
        name: area
        required: true
        schema:
          $ref: '#/definitions/lab_internal_app_DTO.ChangeSolarPanelAreaRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Обновленная панель в заявке
          schema:
            $ref: '#/definitions/lab_internal_app_ds.RequestPanels'
        "400":
          description: Некорректное значение площади
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Не авторизован
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Заявка недоступна этому пользователю
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Заявка или панель не найдены
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Изменить площадь панели в заявке
      tags:
      - RequestPanels
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"